Mostly Study and Documentation
i am just adding comments to explain what i am seeing and how it works so that i actually understand it.  i have just learned that rust provides an easy way to read environment variables. so no more needing a dot env file to store the token. that is an actual innovation compared to previous attempts at making a bot. much more secure that way as it makes it more difficult to leak the key by accident.